routing: redirect / and /trade to fallback ticker
